+++ src/app/components/Feed.tsx
@@ -1,12 +1,4 @@
-import { AiFillStar, AiOutlineUser } from "react-icons/ai"
-
-type CardProps = {
-    title:string,
-    description:string,
-    date:string,
-    likes:number,
-    author:string
-}
+import Card from "./feed/Card"
 
 const Feed = () => {
     const notes = [
@@ -41,32 +33,5 @@ const Feed = () => {
     }
   </div>
   )
-}
-const Card = ({title,description,date,author,likes}:CardProps) => {
-    return (
-        <div className="max-w-2xl p-6 rounded-lg bg-light drop-shadow-md dark:bg-gray-800 dark:border-gray-700 outline-1 outline-black outline">
-            <a href="#">
-                <h5 className="mb-2 text-2xl font-bold tracking-tight text-gray-900 dark:text-white">
-                    {title}
-                </h5>
-            </a>
-            <p className="mb-3 font-normal text-gray-700 dark:text-gray-400">
-                {description}
-            </p>
-            <div className="flex justify-between">
-                <div className="flex items-center gap-3">
-
-                    <AiOutlineUser />
-                    <span>{author}</span>
-                    <span>{date}</span>
-                </div>
-                <div className="flex items-center gap-2">
-                    <AiFillStar />
-                    <span>{likes}</span>
-                </div>
-            </div>
-        </div>
-
-    )
 }
 export default Feed
